% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_df.R
\name{anonymize_df}
\alias{anonymize_df}
\title{anonymize_df}
\usage{
anonymize_df(df, cols_to_anon, algo = "crc32", preserve_NA = TRUE)
}
\arguments{
\item{df}{Dataframe that contains columns you want to anonymize}

\item{cols_to_anon}{Character vector with column names to anonymize}

\item{algo}{The type of algorithm to use}

\item{preserve_NA}{Change to FALSE if you want NA to be anonymized as well}

\item{out}{What to output, i.e. original df, anonymized df, keys, seeds (note that seeds are currently randomy generated)}
}
\value{
A list, a dataframe, or a vector, depening on out
}
\description{
anonymize_df
}
