% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deanonymize_df.R
\name{deanonymize_df}
\alias{deanonymize_df}
\title{deanonymize dataframe}
\usage{
deanonymize_df(anonymized_df, keys, cols_to_deanon = "names(keys)",
  key_col_anon = 3, key_col_val = 1)
}
\arguments{
\item{anonymized_df}{Dataframe with values in specified columns that have been anonymized}

\item{keys}{List of dataframes generated from anonymize_df}

\item{cols_to_deanon}{Default uses names from `keys`, but can specify specific columns}

\item{key_col_anon}{The column position or name in the `keys` dataframes where the current anonymized value is}

\item{key_col_val}{The column position or name in the `keys` dataframes where the original value is}
}
\value{
A dataframe.
}
\description{
deanonymize dataframe
}
